build: Query DSL
