+++ src/main/java/com/project/skb/config/dsl/QueryDslConfig.java
@@ -1,0 +1,21 @@
+package com.project.skb.config.dsl;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+@Configuration
+public class QueryDslConfig {
+
+    @PersistenceContext
+    public EntityManager em;
+
+    @Bean
+    public JPAQueryFactory japQueryFactory(){
+        return new JPAQueryFactory(em);
+    }
+
+}
