+++ src/main/java/com/project/skb/config/security/CustomUserDetailService.java
@@ -1,6 +1,6 @@
-package com.project.skb.config;
+package com.project.skb.config.security;
 
-import com.project.skb.user.domain.UserRepository;
+import com.project.skb.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
