+++ src/main/java/com/project/skb/config/security/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.project.skb.config;
+package com.project.skb.config.security;
 
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.core.Authentication;
