+++ src/main/java/com/project/skb/post/repository/PostRepository.java
@@ -1,9 +1,10 @@
-package com.project.skb.post.domain;
+package com.project.skb.post.repository;
 
+import com.project.skb.post.domain.Post;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
-public interface PostRepository extends JpaRepository<Post,Long> {
+public interface PostRepository extends JpaRepository<Post,Long>, PostRepositoryCustom {
 
     List<Post> findByType(String type, Pageable pageable);
 }
