+++ src/main/java/com/project/skb/post/repository/PostRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.project.skb.post.repository;
+
+import com.project.skb.post.domain.Post;
+
+import java.util.List;
+
+public interface PostRepositoryCustom {
+
+    List<Post> getPosts(int page);
+}
